Move Memory Protection Enforcement to GCD Layer
Part of memory protection management (specifically setting
RP on free memory and XP on newly allocated memory) was occuring
in the memory block layer instead of the GCD layer. This was
inappropriate coupling of the memory protection feature to a more
generic layer and caused the feature to unnecessarily and complexly
span layers. This was necessary at the time because of a different
design, but is no longer required.

This commit moves the memory protections to the GCD layer and out
of the memory block layer.