Allocate Initial Memory Blocks with Page Granularity
The memory block initialization code was not setting UEFI_PAGE_SHIFT
as the required align when calling allocate_memory_space, which could
cause it be non-page aligned, which would be fatal. This was unlikely
because the PHIT HOB should pass in a page aligned base address, but
the memory block code should ensure this.

This commit also updates some tests to use UEFI_PAGE_SHIFT.