uefi_sdk_macro: Prevent alloc name collision between macros (#60)
## Description

If the sole purpose of a component is to produce a service to be used by
other components, the developer must create two structs: (1) the struct
representing the component that gets dispatched and (2) the struct
representing the service.

In reality, you can actually create a single struct that represents both
the component and the service. This allows the struct to be instantiated
and registered as a component, then during component execution, it
registers itself as a service.

This works, however a name collision for aliasing the alloc crate
prevents the `IntoService` and `IntoComponent` macro from being derived
on the same object. This PR fixes that.

```rust
use uefi_sdk::error::Result;
use uefi_sdk::component::Storage;
use uefi_sdk::component::service::IntoService;
use uefi_sdk::component::IntoComponent;
use uefi_sdk::component::service::Service;

trait MyTrait {
    fn do_something(&self);
}

trait OtherTrait {
    fn do_something_else(&self);
}

#[derive(IntoComponent, IntoService, Default)]
#[service(dyn MyTrait)]
struct ComponentAndService {
    inner: Option<Service<dyn OtherTrait>>,
}

impl MyTrait for ComponentAndService {
    fn do_something(&self) {
        if let Some(service) = &self.inner {
            service.do_something_else();
        }
        log::info!("ComponentAndService::do_something called!");
    }
}

impl ComponentAndService {
    fn entry_point(mut self, storage: &mut Storage, service: Service<dyn OtherTrait>) -> Result<()> {
        log::info!("ComponentAndService entry point called!");
        self.inner = Some(service);
        storage.add_service(self);
        Ok(())
    }
}
```

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated that the above example compiles and it can be registered with
the core.

## Integration Instructions

<_Describe how these changes should be integrated. Use N/A if nothing is
required._>