component: Clarify conflicting parameters (#64)
## Description

Currently a Storage parameter is referred to as a Config or ConfigMut
depending on the context. Since it is likely as components are first
being created conflicts may arise, try to make the message more precise
by indicating if the previous parameter is likely Storage.

My understanding is that this is mostly accurate now but may become less
so if new parameters are added in the future and then a more scalable
solution might be needed.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- Unit tests
- Execute component in QEMU boot

## Integration Instructions

- N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>