Improve Boot Services Protocol Interface (#38)
## Description

improve install_protocol_interface ergonomics
https://github.com/OpenDevicePartnership/uefi-sdk/issues/34

To fix most of these issues, making the `Protocol` trait implementation
would help fix some issue and make the use of it clearer. In this
implementation we also don't need to deref as a guid because now the
struct that implement `Protocol` do not represent a guid and if we need
it, we can do `MyProtocolInterface::PROTOCOL_GUID`.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [X] Breaking change?
- [X] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Unit tests
Have been used in DxeRustHidV2

## Integration Instructions

N/A

---------

Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>