[RFC] dependency injectable Guided HOB parsing (#349)
RFC for adding a new dependency injectable type `Hob`, which when found
in a registered component, also registers a parser for said HOB. This parser
will be used to parse a guided HOB into the `Hob<T>` type that the
component is requesting. The component will only be executed if the HOB
is found.