uefi_sdk_macro: Fix HOB macro GUID conversion (#77)
## Description

The `FromHob` trait derivation is still not working as expected for me
after ffa097d7efee8b1204fee58379180f5a0753cb48.

The `time_low`, `time_mid`, and `time_hi_and_version` fields do not
match what is actually in the HOB list.

For the HOB string: "`d4ffc718-fb82-4274-9afc-aa8b1eef5293`":

Actual (from HOB list):
```
  { time_low: [24, 199, 255, 212], time_mid: [130, 251],
    time_hi_and_version: [116, 66], clk_seq_hi_res: 154,
	clk_seq_low: 252, node: [170, 139, 30, 239, 82, 147] }
```

Value in HOB parser list:
```
  { time_low: [212, 255, 199, 24], time_mid: [251, 130],
    time_hi_and_version: [66, 116], clk_seq_hi_res: 154,
	clk_seq_low: 252, node: [170, 139, 30, 239, 82, 147] }
```

Fields after this are as expected.

A well-known HOB GUID for the PCD Database HOB was added as an unit test
to show what the `hob` string is expected to be and that the hex value
(which is easy to compare to MdeModulePkg.dec) translates to the
expected string value.

This also matches the HOB I expect correctly.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- Unit tests
- Verify HOB expected in HOB list is found with expected `hob` string
value
- Compare generated macro value versus equivalent GUID value dumped from
HOB list

## Integration Instructions

N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>