Service: Implement Copy + Clone (#80)
## Description

As services are always immutable (they have interior immutability where
necessary), it is inherently safe to allow copy and cloning of services.
This is useful for scenarios where two instances are necessary, such as
producing two services, both of which utilize the same underlying
service. This does not actually clone anything, just creates a new
instance of the service struct pointing at the same underlying leaked
data.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Added unit tests

## Integration Instructions

N/A