component: Storage: Deprecate public try_* functions (#79)
## Description

in core/standard rust, try_* functions are typically associated with
fallable operations that return a Result. get_* functions are used for
non-fallable operations that return an Option (See any rust container
or collection such as Vec, HashMap, etc). To be consistent with this
convention, we are replacing the public try_get_* functions in the
Storage component with the get_* functions

- [ ] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

Replace any usage of `try_get_*` with `get_*`. The interface and
implementation did not change.