config: Do not always instantiate the default implementation (#86)
## Description

`Config` / `ConfigMut` requires that the underlying type implement
`Default`. There is a valid use case that the `Default` implementation
is never acceptable and must be provided directly to the `Core`. In this
scenario, it makes sense to be allowed to panic inside of the `Default`
trait implementation, Informing the Developer that they must provide a
value.

Currently, this is not possible because the default implementation is
always generated, regardless of if an instance is provided to the Core.
Not only does this enable the use case of panicking if the `Default`
implementation is used, but it also fixes a performance issue where we
were instantiating the default value of a config when it was not
necessary.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated the `default()` is not called unless necessary

## Integration Instructions

N/A