component: Add Commands Param type (#95)
## Description

Adds a new parameter, `Commands` that can be used to apply structural
changes to the world without causing a `Param` conflicts on types that
depend on `Storage` not being manipulated (right now just `Config` and
`ConfigMut`). This is done by not actually applying the commands
immediately, but instead waiting to apply the commands until after the
component executes.

As it stands right now, since we do not run in parallel, this means any
commands will be applied immediately after the component is executed
(truthfully it is applied right before the next component is executed).
However, in the future, if we go parallel, the commands would not be
executed until there is a period of time in which only a single
component is about to run (Either the start of the next iteration of
components, or we have collapsed execution to a single component for
some reason).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [x] Includes documentation?

## How This Was Tested

unit tests

## Integration Instructions

N/A - users may now use the `Commands` parameter.