dxe_core: Update init_paging to Sync NX to GCD When Applying to Paging (#388)
## Description

Currently, part of the init_paging function uses GCD APIs to manipulate
page table attributes in the newly built page table, but not for already
allocated memory, MMIO, or reserved memory. This creates a mismatched
state in the GCD (NX not reflected) and page table (NX reflected).

This patch updates init_paging to exclusively use the GCD APIs to
manipulate the page table (even the one being formed) so that the GCD
always stays in sync. This matches all other updates of the page table.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested booting Q35 and SBSA to Windows and Linux and breaking in with
the debugger, dumping the GCD to confirm all regions expected had NX set
and tested a few of these regions in the page table to confirm NX was
set.

## Integration Instructions

N/A.