Disable cargo release dry run - To pass CI for patina monorepo PR (#400)
## Description

Disable Cargo Release Dry Run - To Pass CI for Patina Monorepo PR

[https://github.com/OpenDevicePartnership/uefi-dxe-core/pull/399](https://github.com/OpenDevicePartnership/uefi-dxe-core/pull/399)

**Why do we need this(one off exception)?**
With the above Patina PR, we are consolidating code from three different
repositories (`uefi-dxe-core`, `uefi-core`, and `uefi-sdk`) into a
single versioned repository for the Patina project. As part of this
consolidation, we have selected `15.0.0` as the next version for the
Patina repository.

All crates now inherit the version from the Patina workspace
`Cargo.toml`, and dependencies are declared as shown below:

```toml
# Cargo.toml
# sdk/boot_services dependencies
# -----------------------------------
uefi_protocol = { version = "15.0.0", path = "sdk/uefi_protocol" }
```
Currently, our CI setup for any PR performs a dry run of the release
process, which includes checking whether the specified dependency
versions are available in the internal registry. Since version `15.0.0`
has not yet been published, the dry run fails.

This change disables the dry run in order to allow the CI to pass for
the PR mentioned above. Once the PR is merged, we will follow the usual
process of generating a draft release, and the `publishrelease` workflow
will handle the actual publication of the crates.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

NA

## Integration Instructions

NA

---------

Signed-off-by: Vineel Kovvuri[MSFT] <vineelko@microsoft.com>