service: Store services as static reference
Updates the Service struct to expect the downcast target to be a static
reference rather than a Box, and updates the IntoService macro to
generate the downcast target as a static reference.

This change is to support the ability to register a concrete type that
is already a static reference, such as a struct that is instantiated
at compile time via the `static` keyword.