Make copy of device path in LoadImage for loaded image protocol (#429)
This change makes a copy of the device path in the `LoadImage` function for the loaded image protocol. This is not explicitly state for loaded_image_protocol, but is mentioned for the required by the UEFI specification for `EFI_LOADED_IMAGE_DEVICE_PATH_PROTOCOL`.

This specifically revolves a use after free issue that presented in the shell which happend to overwrite a device path when it was re-allocated this memory while the pointer was still in the loaded image protocol. Whenever this protocol is next inspected it can cause unexpected behavior. Specifically, the `dh` command would cause the device to get stuck in a loop trying to parse the invalid path and eventually assert.