uefi_sdk: Add service benchmarks
Adds two benchmarks for the Service Parameter type to measure the
performance costs of executing a service. Specifically this is looking
comparing the performance costs of using a dyn service vs a concrete
service type (Service<dyn MyService> vs Service<MyServiceImpl>).

This can also be used to measure the performance differences of service
execution in general when making changes to the Service parameter type.