uefi_sdk: support Service const instantiation
This supports the Service struct to be instantiation within a const
context (albeit in an uninitialized state). This is useful for when a
service is needed in a const context, such as when creating a static
instance of something that uses a service.

This also implements Send + Sync for the Service struct, which is
possible because all access to services is done through a Deref to the
underlying type and thus does not allow any mutable access. Services
are required to implement their own interior mutability.