Make performance lib into a component. (#405)
## Description

Change the performance library into a component, refactor parts of the
code for it to be more testable and add unit tests.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [X] Includes tests?
- [X] Includes documentation?

## How This Was Tested
- Unit tests
- On QEMU Q35
- Intel devkit

## Integration Instructions

This Performance component replaced DxeCorePerformanceLib.c. This means
that if the device is not built using performance, this component should
not be added and if it is built with performance, it should be there.
This is because the component publishes the fbpt but when perf is not
enabled, FirmwarePerformanceDxe.c creates that table. This is the PCD
that tells whether perf is enabled or not:
gEfiMdePkgTokenSpaceGuid.PcdPerformanceLibraryPropertyMask. Note that in
rust the value of the pcd is not used and to enable some measurement,
config must be used.

---------

Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>