Dispatcher: Reduce Depex Evaluation Message to Trace
One of the noisest logs in the patina boot currently is the
log that says a depex guid was evaluated, which in general does
not provide useful information to an average build. This commit
moves that to a trace log.

It also moves all other depex info logs to trace, as they are not
generally useful outside of deep depex debugging.