Image Loader: Remove Runtime Image Memory Protections on EBS
edk2 follows a pattern where runtime image memory protections
are removed on ExitBootServices to allow the SetVirtualAddressMap
callback to relocate runtime images. Patina was not following this,
which only matters for certain flavors of Linux, it could cause
the core to attempt to write to RO memory.

For now, Patina follows the same method as edk2, but investigation
is happening to determine if we should have a different method.