Support for Hoblist with dxe core in allocation hob (#402)
Fixes https://github.com/OpenDevicePartnership/patina/issues/323 
This is already reviewed code and was asked to move this code as part of
another PR as per Michael. Hence raising another PR for this.

Also - this addresses the comments given from
https://github.com/OpenDevicePartnership/uefi-dxe-core/pull/326.

The current hoblist in testsupport crate is not sufficient and does not
contain the dxe-core HOB with dxe core efi filled in expected by our
initiatlisation. When existing hoblist is used, it panics.
Hence there was a need to create another hoblist with dxe-core and fill
in with dxe EFI.
Based on comments from
https://github.com/OpenDevicePartnership/uefi-dxe-core/pull/326.
The comment mentioned on taking only the hoblist part and push to main.
Hence creating this issue for tracking.

---------

Co-authored-by: Rathi Kartheek <rkartheek@microsoft.com>