EventDb: Allow re-entrance to tpl_restore from event notify functions. (#446)
## Description

The present implementation of `restore_tpl` uses an atomic flag
`EVENT_NOTIFIES_IN_PROGRESS` to handle the issue where locking the EVENT
DB triggers a `restore_tpl` that would result in an endless loop. The
implementation currently holds that flag while invoking the event
notification function which has the effect of preventing notification
callbacks _during_ the execution of event notifies.

Generally this is not an issue, but there are some corner cases where
this can violate expectations of code that expects to manipulate the TPL
within the event callback to trigger other event notifies. An example of
where this is in the UsbBusDxe driver here:
https://github.com/microsoft/mu_basecore/blob/3fd74e217a1e6c72f90efb1517f7da610c42d074/MdeModulePkg/Bus/Usb/UsbBusDxe/UsbHub.c#L1009

This while loop will hang unless timer events fire while it is looping -
and the `EVENT_NOTIFIES_IN_PROGRESS` implementation presently prevents
that.

This PR reworks the `restore_tpl` implementation such that the
`EVENT_NOTIFIES_IN_PROGRESS` flag is released before executing the
notification functions - this allows re-entrancy to `restore_tpl` from
the notification functions while still addressing the infinite regress
issue from the event db lock.

In addition, this PR removes the EventNotificationIterator object as it
is no longer required given the rework of the code. An equivalent
iterator can be generated with the `core::iter::from_fn` function (as is
demo'd in the updated test code).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Basic functional testing to ensure that pathological case from UsbBusDxe
above is addressed functionally, updated unit tests

## Integration Instructions

N/A - no public APIs changed.