Move GCD interaction from FixedSizeBlockAllocator to SpinLockedFixedSizeBlockAllocator (#435)
## Description

To avoid interacting with the GCD in the FSB's locked regions (which can
potentially cause an FSB lock re-entrancy if the GCD allocates memory),
this change moves all interaction with the GCD outside of the locked
regions of the FSB and into the spin-locked FSB. As such, new memory
allocations from the GCD are now either requested by the FSB then
allocated from the spin-locked FSB (in the case of alloc/dealloc) or
made by the spin-locked FSB, which then informs the FSB of the change.

Main changes include:
- Rather than expanding by allocating memory from the GCD directly, the
FSB's `expand` is now passed-in a region of memory to expand itself with
which is allocated from the GCD by the spin-locked FSB
- Attempted allocations in the FSB will return
`FixedSizeBlockAllocatorError::OutOfMemory(additional_mem_requested)` if
the FSB is unable to fit the allocation, prompting the spin-locked FSB
to allocate additional memory with the GCD and call `expand` on the FSB
- Note: This has the side effect that the FSB's `pool_allocation_calls`
stat would be incremented twice in the case that additional memory needs
to be allocated by the spin-locked FSB
- As there is no longer a direct analogue for the `alloc`/`dealloc`
functions of `GlobalAlloc` and the `allocate`/`deallocate` functions of
`Allocator`, the FSB now defines only one `alloc` and `dealloc` which
are shared by both trait implementations
- The bulk of the allocation logic for `allocate_pages` and `free_pages`
have been moved into the spin-locked FSB, which then notifies the FSB of
any page allocations made/freed with `notify_page_allocation` and
`notify_pages_freed`
- `preferred_range` has been renamed to `reserved_range` as the FSB
itself makes no preference for putting new allocations into the
`preferred_range`, and any preference the GCD would show for allocating
memory in that region would be solely predicated by the
"`preferred_range`" being allocated earlier in the boot process
- GCD interaction as part of `reserve_memory_region` has been moved to
the spin-locked FSB, with the reserved memory region being registered in
the FSB with `set_reserved_range(range)`. Rather than calling into
`allocate_pages` then `free_pages` to reserve the memory region,
`reserve_memory_region` now reserves memory in the GCD by directly
calling `gcd.allocate_memory_space` then
`gcd.free_memory_space_preserving_ownership` to avoid having to undo the
stats side-effects of the `allocate_pages` and `free_pages` functions.
- Note: This has the effect of not iterating `page_allocation_calls` and
`page_free_calls` when calling `reserve_memory_pages`
- To keep things consistent, raw pointers and raw pointers + usize
lengths have been largely replaced with `NonNull<u8>` and
`NonNull<[u8]>` respectively.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Booted to Windows on x64 physical platform.

## Integration Instructions

N/A