Debugger: Invalidate TLB and instruction caches on exit (#414)
The debugger may edit the page table mappings or instructions in the
instruction cache. These caches are not inherently coherent with the
data cache and so these should be invalidated on returning from the
debugger.

Notably, the x64 memory model does guarantee writes will invalidate the
icache so long as they use the same linear address, which will be true
since we only operate identity mapped.