Update core_disconnect_controller to not disconnect drivers that are already disconnected and preserve ordering. (#470)
## Description

`core_disconnect_controller` identifies all drivers that have any
protocol on the given controller handle open by driver and attempts to
stop them all. In some cases, however, stopping a driver earlier in the
list may have the side effect of stopping a driver not yet processed.
Prior to this change, a driver that had already been stopped was
unconditionally processed again when it was reached in the list,
resulting in a second call to it driver_binding->stop(). This is
unexpected behavior not aligned with the EDK2 C core reference or the
spec.

This change ensures that a driver that has already been stopped is
skipped when it is later processed. In addition, it updates the driver
and child handle processing to maintain stable ordering (i.e. driver and
child handles are processed in the order they were added to the handle).

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Functionally tested using USB hot add/hot remove flows to stimulate
relatively complicated connect/disconnect flows. Observed no more double
stops() with this change.

## Integration Instructions

N/A