patina_sdk:memory: Implement mock MemoryManager (#448)
## Description

Implements a working MemoryManager implementation for the MemoryManager
trait by using the global allocator. This should be used when a mock is
needed for testing purposes, that can allocate and deallocate memory as
expected.

This commit also adds mockall's automock to the MemoryManager trait, to
simplify mocking MemoryManager interface method failures in tests.

Overall this provides two mocks for MemoryManager automock's
`MockMemoryManager` and a custom working MemoryManager,
`StdMemoryManager`.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [x] Includes documentation?

## How This Was Tested

Compiles

## Integration Instructions

N/A