Handle special interrupt IDs for hardware interrupt handler (#494)
## Description

Current implementation did not take into account the special interrupt
IDs (1021, 1022, and 1023). Thus when such interrupts occur, the
`raw_value` will be used as an index into the handler vector, causing a
panic.

This change is to handle the special interrupt IDs by checking if the
`raw_value` is a special ID, otherwise it needs to be within the range
of the handler vector.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on physcial hardware and booted to Windows.

## Integration Instructions

N/A