Fix install_configuation_table API
install_configuration_table took in a &c_void which is meaningless
to hold a reference to instead of just having a *c_void. The former
approach actually requires more unsafe code from callers to arrange
the *c_void they have into a &c_void.

Additionally, this API was taking an Option to this when really it
wanted to check if the pointer was null or not, as per UEFI spec,
so that is updated.