Add Patina Requirements Documentation (#521)
## Description

This adds Patina Requirements Documentation to document the differences
between edk2 and Patina and give justification and guidance on how
platforms can meet these requirements.

This documentation is written from the perspective of describing
Patina's handoff requirements with guidance on how to meet them, but an
additional document can be written that is targeted directly at a
platform integrator that delves into more detail, if that is deemed
necessary.

The page was renamed from rust vs edk2 to Patina Requirements, to have
Patina stand on its own and not just be seen as a derivative of edk2,
while still calling out the differences from edk2, as that is where
platforms will be transitioning from. This can be changed if folks
prefer.

When the Patina DXE Readiness Tool moves to GitHub, the doc will be
updated to link to it.

Closes #516.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

## How This Was Tested

N/A.

## Integration Instructions

N/A.