Replace Uuid->Guid transmute with safe equivalent (#532)
## Description

`r_efi::Guid::from_bytes` is implemented with the same transmute so the
generated code is expected to be identical to the original however the
use of `unsafe` is now encapsulated.

This removes a significant number of uses of `mem::transmute` from the
workspace.

Noticed while reviewing some of the code in the workspace.
 
- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Built locally without error.

## Integration Instructions

N/a