patina-sdk: Make align* Functions More User Friendly
Currently the align family of functions in the sdk requires
u64s to be passed in, which often requires a pattern of
my_usize as u64 getting passed in and the result of this
being cast back to usize. This is needless, these functions
can easily be generic.

In addition to this, these functions were passing back string
errors which were universally being converted to EfiErrors. It
was not using to return strings, instead return EfiErrors.

While we are here, add some unit tests for these functions.