Add device path. (#407)
## Description

This PR adds an abstraction over device path so we can do operations on
them natively in rust. DevicePathBuf is an own version of the type and
DevicePath is used as a borrow version of it. Similar to Path and
PathBuf in std rust. The memory layout in the internal vector is the
same as the one in C, so these structs can be easily converted and used
from both languages.

This feature is unstable for now, enable feature `unstable-device-path`
to use it:
https://github.com/OpenDevicePartnership/patina/issues/547


- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [X] Includes tests?
- [X] Includes documentation?

## How This Was Tested

Converted a device path pointer from C and then printed it to see if the
layout was read correctly.

## Integration Instructions

N/A

---------

Co-authored-by: John Schock <joschock@microsoft.com>
Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>